refactor: divide package 'common' into several smaller packages
