+++ src/main/kotlin/org/rucca/cheese/common/error/BaseError.kt
@@ -1,4 +1,4 @@
-package org.rucca.cheese.common
+package org.rucca.cheese.common.error
 
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.databind.JsonSerializer
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
 import org.springframework.http.HttpStatus
 
-private class ErrorSerializer : JsonSerializer<org.rucca.cheese.common.BaseError>() {
-    override fun serialize(err: org.rucca.cheese.common.BaseError, gen: JsonGenerator, serializer: SerializerProvider) {
+private class ErrorSerializer : JsonSerializer<BaseError>() {
+    override fun serialize(err: BaseError, gen: JsonGenerator, serializer: SerializerProvider) {
         gen.writeStartObject()
         gen.writeNumberField("code", err.status.value())
         gen.writeStringField("message", "${err.name}: ${err.message}")
@@ -23,7 +23,7 @@ private class ErrorSerializer : JsonSerializer<org.rucca.cheese.common.BaseError
     }
 }
 
-@JsonSerialize(using = org.rucca.cheese.common.ErrorSerializer::class)
+@JsonSerialize(using = ErrorSerializer::class)
 abstract class BaseError(
         val status: HttpStatus,
         val name: String,
