+++ src/main/kotlin/org/rucca/cheese/common/error/InternalServerError.kt
@@ -1,9 +1,9 @@
-package org.rucca.cheese.common
+package org.rucca.cheese.common.error
 
 import org.springframework.http.HttpStatus
 
 class InternalServerError :
-        org.rucca.cheese.common.BaseError(
+        BaseError(
                 HttpStatus.INTERNAL_SERVER_ERROR,
                 "InternalServerError",
                 "The server encountered an unexpected error. Please try again later.")
