+++ src/main/kotlin/org/rucca/cheese/common/error/NotFoundError.kt
@@ -1,12 +1,13 @@
-package org.rucca.cheese.common
+package org.rucca.cheese.common.error
 
+import org.rucca.cheese.common.persistent.IdType
 import org.springframework.http.HttpStatus
 
 class NotFoundError(
         type: String,
-        id: org.rucca.cheese.common.IdType,
+        id: IdType,
 ) :
-        org.rucca.cheese.common.BaseError(
+        BaseError(
                 status = HttpStatus.NOT_FOUND,
                 name = "NotFoundError",
                 message = "$type with id $id was not found",
